Stuck...
I'm stuck with rcConnect() for now and I have no idea on how to proceed. So I have written some to-dos and explanations as comments. I will continue with other parts and mimic iput, icd and other basic iCommands. The general idea for now is: as long it compiles, it has to be right... :)